fix(fix-llvm): apply shellcheck

+++ .github/workflows/fix-llvm-header-guard.sh
@@ -1,24 +1,29 @@
+#!/usr/bin/env bash
 while IFS= read -r line
 do
-    echo "$line" | grep -q "llvm"
-    if [ $? -eq 0 ]; then
-        IFS= read "actual"
-        IFS= read "tild"
-        IFS= read "expected"
-        actualFiltered=$(echo $actual | grep -oE '[^ ]+$')
-        expectedFiltered=$(echo $expected | grep -oE 'CHESS.*$')
+
+    if echo "$line" | grep -q "llvm" ; then
+        IFS= read -r "actual"
+        IFS= read -r "tild"
+        IFS= read -r "expected"
+        actualFiltered=$(echo "$actual" | grep -oE '[^ ]+$')
+        expectedFiltered=$(echo "$expected" | grep -oE 'CHESS.*$')
 
         if [ "$actualFiltered" = "$expectedFiltered" ]; then
             continue
         else
-            echo "$line" >> clang-tidy-tmp
-            echo "$actual" >> clang-tidy-tmp
-            echo "$tild" >> clang-tidy-tmp
-            echo "$expected" | sed 's/[^ ]*CHESS/CHESS/g' >> clang-tidy-tmp
+            {
+                echo "$line"
+                echo "$actual"
+                echo "$tild"
+                echo "$expected" | sed 's/[^ ]*CHESS/CHESS/g'
+            } >> clang-tidy-tmp
         fi
     else
-        echo "$line"
+        echo "$line" >> clang-tidy-tmp
     fi
 done < clang-tidy-output
 
-mv clang-tidy-tmp clang-tidy-output
+if [ -f clang-tidy-tmp ]; then
+  mv clang-tidy-tmp clang-tidy-output
+fi
